backend/requirements.txt
# Core FastAPI and Server
fastapi[standard]==0.116.1
uvicorn[standard]==0.30.6
websockets==13.1

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.22
alembic==1.11.1

# Configuration and Settings
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.11.0

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Google Cloud ASR
google-cloud-speech==2.27.0
google-auth==2.35.0

# Async File Handling
aiofiles==24.1.0